refactor: extract shared layout components and hooks
- Extract PageLayout component (replaces MenuProvider+Header+CircleMenu+Footer boilerplate in 10 pages)
- Extract AdminLayout as Next.js layout.tsx (replaces auth guard + sidebar in 7 admin pages)
- Extract Wave animation component (shared between Stats and StatsSkeleton)
- Extract useNewsletterSubscription hook (shared between Newsletter and Cta)
- Extract useSeasonFallback hook (shared between AllWedgiesPage and StandingsPage)
