+++ src/app/admin/layout.tsx
@@ -1,0 +1,36 @@
+import { auth } from "~/server/auth";
+import { HydrateClient } from "~/trpc/server";
+import { SignIn } from "~/components/admin/auth";
+import { SidebarProvider } from "~/components/ui/sidebar";
+import { AppSidebar } from "~/components/sidebar";
+import { AdminHeader } from "~/components/admin/header";
+
+export default async function AdminLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  const session = await auth();
+
+  if (!session) {
+    return (
+      <main className="flex min-h-screen items-center justify-center bg-gray-100">
+        <div className="text-center">
+          <SignIn />
+        </div>
+      </main>
+    );
+  }
+
+  return (
+    <HydrateClient>
+      <SidebarProvider defaultOpen={false}>
+        <AppSidebar />
+        <main className="min-h-screen w-full bg-darkpurple">
+          <AdminHeader />
+          <div className="container mx-auto p-6">{children}</div>
+        </main>
+      </SidebarProvider>
+    </HydrateClient>
+  );
+}
