+++ src/app/admin/page.tsx
@@ -1,10 +1,4 @@
-import { auth } from "~/server/auth";
-import { api, HydrateClient } from "~/trpc/server";
-import { SignIn } from "~/components/admin/auth";
-
-import { SidebarProvider } from "~/components/ui/sidebar";
-import { AppSidebar } from "~/components/sidebar";
-import { AdminHeader } from "~/components/admin/header";
+import { api } from "~/trpc/server";
 import { AdminDash } from "~/components/admin/dash";
 import { generateMetadata } from "~/config/metadata";
 
@@ -14,32 +8,7 @@ export const metadata = generateMetadata({
 });
 
 export default async function Home() {
-  const session = await auth();
+  void api.admin.getGlobal.prefetch();
 
-  if (session?.user) {
-    void api.admin.getGlobal.prefetch();
-  }
-
-  if (!session) {
-    return (
-      <main className="flex min-h-screen items-center justify-center bg-gray-100">
-        <div className="text-center">
-          <SignIn />
-        </div>
-      </main>
-    );
-  }
-
-  return (
-    <HydrateClient>
-      <SidebarProvider defaultOpen={false}>
-        <AppSidebar />
-        <main className="min-h-screen w-full bg-darkpurple">
-          <AdminHeader />
-
-          <AdminDash />
-        </main>
-      </SidebarProvider>
-    </HydrateClient>
-  );
+  return <AdminDash />;
 }
