+++ src/app/admin/wedgies/[id]/page.tsx
@@ -1,6 +1,4 @@
-import { auth } from "~/server/auth";
 import { api } from "~/trpc/server";
-import { SignIn } from "~/components/admin/auth";
 import { WedgieFormPage } from "~/components/admin/WedgieFormPage";
 import { notFound } from "next/navigation";
 
@@ -11,27 +9,12 @@ interface PageProps {
 }
 
 export default async function EditWedgiePage({ params }: PageProps) {
-  const session = await auth();
   const { id } = await params;
-
-  if (!session) {
-    return (
-      <main className="flex min-h-screen items-center justify-center bg-gray-100">
-        <div className="text-center">
-          <SignIn />
-        </div>
-      </main>
-    );
-  }
 
   // If id is "new", we're creating a new wedgie
   if (id === "new") {
     const global = await api.admin.getGlobal();
-    return (
-      <main className="min-h-screen w-full bg-darkpurple p-6">
-        <WedgieFormPage currentSeason={global?.currentSeason?.name} />
-      </main>
-    );
+    return <WedgieFormPage currentSeason={global?.currentSeason?.name} />;
   }
 
   // Otherwise, fetch the existing wedgie
@@ -40,9 +23,5 @@ export default async function EditWedgiePage({ params }: PageProps) {
     return notFound();
   }
 
-  return (
-    <main className="min-h-screen w-full bg-darkpurple p-6">
-      <WedgieFormPage wedgie={wedgie} />
-    </main>
-  );
+  return <WedgieFormPage wedgie={wedgie} />;
 }
