+++ src/app/blog/page.tsx
@@ -1,10 +1,7 @@
 export const dynamic = "force-static";
 export const revalidate = 3600;
 
-import { Header } from "~/components/layout/Header";
-import { CircleMenu } from "~/components/layout/CircleMenu";
-import { MenuProvider } from "~/context/MenuContext";
-import { Footer } from "~/components/layout/Footer";
+import { PageLayout } from "~/components/layout/PageLayout";
 import { BlogList } from "~/components/blog/BlogList";
 import { generateMetadata } from "~/config/metadata";
 
@@ -15,27 +12,22 @@ export const metadata = generateMetadata({
 
 export default function BlogPage() {
   return (
-    <MenuProvider>
-      <div className="flex min-h-screen flex-col bg-darkpurple">
-        <Header />
-        <CircleMenu />
-        <div className="flex flex-col">
-          <div className="flex w-full flex-col items-center justify-center gap-8 px-4 py-4 md:gap-8 md:py-8 lg:px-8 lg:py-8">
-            <h1 className="text-center text-4xl font-black uppercase leading-none md:text-6xl">
-              <span className="text-shadow-darkpurple relative z-10 block leading-none text-yellow">
-                Wedgie
-              </span>
-              <span className="relative z-0 mt-[-.3em] block text-[1.4em] leading-none text-pink">
-                Blog
-              </span>
-            </h1>
-            <div className="min-h-[60vh] w-full">
-              <BlogList />
-            </div>
+    <PageLayout>
+      <div className="flex flex-col">
+        <div className="flex w-full flex-col items-center justify-center gap-8 px-4 py-4 md:gap-8 md:py-8 lg:px-8 lg:py-8">
+          <h1 className="text-center text-4xl font-black uppercase leading-none md:text-6xl">
+            <span className="text-shadow-darkpurple relative z-10 block leading-none text-yellow">
+              Wedgie
+            </span>
+            <span className="relative z-0 mt-[-.3em] block text-[1.4em] leading-none text-pink">
+              Blog
+            </span>
+          </h1>
+          <div className="min-h-[60vh] w-full">
+            <BlogList />
           </div>
         </div>
-        <Footer />
       </div>
-    </MenuProvider>
+    </PageLayout>
   );
 }
