+++ src/app/standings/page.tsx
@@ -2,11 +2,7 @@ export const dynamic = "force-static";
 export const revalidate = 120;
 
 import { StandingsPage } from "~/components/standings/StandingsPage";
-import { Header } from "~/components/layout/Header";
-import { CircleMenu } from "~/components/layout/CircleMenu";
-import { MenuProvider } from "~/context/MenuContext";
-import { Footer } from "~/components/layout/Footer";
-
+import { PageLayout } from "~/components/layout/PageLayout";
 import { generateMetadata } from "~/config/metadata";
 
 export const metadata = generateMetadata({
@@ -17,26 +13,21 @@ export const metadata = generateMetadata({
 
 export default async function Page() {
   return (
-    <MenuProvider>
-      <div className="flex min-h-screen flex-col bg-darkpurple">
-        <Header />
-        <CircleMenu />
-        <div className="flex flex-col lg:flex-row">
-          <div className="flex w-full flex-col items-center justify-center gap-8 px-4 py-4 md:gap-8 md:py-8 lg:px-8 lg:py-8">
-            <h1 className="text-center text-2xl font-black uppercase leading-none md:text-4xl">
-              <span className="text-shadow-darkpurple relative z-10 block leading-none text-yellow">
-                Players/Teams
-              </span>
-              <span className="relative z-0 mt-[-.25em] block text-[1.5em] leading-none text-pink">
-                Standings
-              </span>
-            </h1>
+    <PageLayout>
+      <div className="flex flex-col lg:flex-row">
+        <div className="flex w-full flex-col items-center justify-center gap-8 px-4 py-4 md:gap-8 md:py-8 lg:px-8 lg:py-8">
+          <h1 className="text-center text-2xl font-black uppercase leading-none md:text-4xl">
+            <span className="text-shadow-darkpurple relative z-10 block leading-none text-yellow">
+              Players/Teams
+            </span>
+            <span className="relative z-0 mt-[-.25em] block text-[1.5em] leading-none text-pink">
+              Standings
+            </span>
+          </h1>
 
-            <StandingsPage />
-          </div>
+          <StandingsPage />
         </div>
       </div>
-      <Footer />
-    </MenuProvider>
+    </PageLayout>
   );
 }
