+++ src/app/stats-for-nerds/page.tsx
@@ -3,12 +3,8 @@ export const revalidate = 300;
 
 import { Suspense } from "react";
 import { api } from "~/trpc/server";
-import { Header } from "~/components/layout/Header";
-import { CircleMenu } from "~/components/layout/CircleMenu";
-import { MenuProvider } from "~/context/MenuContext";
-import { Footer } from "~/components/layout/Footer";
+import { PageLayout } from "~/components/layout/PageLayout";
 import { StatsForNerds } from "~/components/stats-for-nerds/StatsForNerds";
-// import { SeasonalProgressChart } from "~/components/stats-for-nerds/SeasonalProgressChart";
 import { Loader } from "~/components/loader";
 import { generateMetadata } from "~/config/metadata";
 
@@ -20,33 +16,24 @@ export const metadata = generateMetadata({
 
 export default async function StatsForNerdsPage() {
   return (
-    <MenuProvider>
-      <div className="flex min-h-screen flex-col bg-darkpurple">
-        <Header />
-        <CircleMenu />
-        <div className="w-full">
-          <div className="flex w-full flex-col items-center justify-center gap-8 px-4 py-4 md:gap-8 md:py-8 lg:px-8 lg:py-8">
-            <h1 className="text-center text-6xl font-black uppercase leading-none md:text-7xl">
-              <span className="text-shadow-darkpurple relative z-10 block leading-none text-yellow">
-                Stats
-              </span>
-              <span className="relative z-0 mt-[-.4em] block text-[.6em] leading-none text-pink">
-                For Nerds
-              </span>
-            </h1>
+    <PageLayout>
+      <div className="w-full">
+        <div className="flex w-full flex-col items-center justify-center gap-8 px-4 py-4 md:gap-8 md:py-8 lg:px-8 lg:py-8">
+          <h1 className="text-center text-6xl font-black uppercase leading-none md:text-7xl">
+            <span className="text-shadow-darkpurple relative z-10 block leading-none text-yellow">
+              Stats
+            </span>
+            <span className="relative z-0 mt-[-.4em] block text-[.6em] leading-none text-pink">
+              For Nerds
+            </span>
+          </h1>
 
-            <Suspense fallback={<LoaderWrapper />}>
-              <StatsWrapper />
-            </Suspense>
-
-            {/* <Suspense fallback={<LoaderWrapper />}>
-              <SeasonalProgressChart />
-            </Suspense> */}
-          </div>
+          <Suspense fallback={<LoaderWrapper />}>
+            <StatsWrapper />
+          </Suspense>
         </div>
-        <Footer />
       </div>
-    </MenuProvider>
+    </PageLayout>
   );
 }
