+++ src/components/Cta.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import React, { useState } from "react";
-import { useToast } from "~/hooks/use-toast";
+import React from "react";
+import { useNewsletterSubscription } from "~/hooks/use-newsletter";
 
 interface CtaProps {
   links: {
@@ -12,43 +12,8 @@ interface CtaProps {
 }
 
 export function Cta({ links, variant = "large" }: CtaProps) {
-  const [email, setEmail] = useState("");
-  const [loading, setLoading] = useState(false);
-  const { toast } = useToast();
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setLoading(true);
-
-    try {
-      const response = await fetch("/api/subscribe", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ email }),
-      });
-
-      const data = (await response.json()) as { error?: string };
-
-      if (response.ok) {
-        toast({
-          title: "Success!",
-          description: "You've been subscribed to our newsletter.",
-        });
-        setEmail("");
-      } else {
-        throw new Error(data.error ?? "Something went wrong");
-      }
-    } catch (error) {
-      toast({
-        title: "Error",
-        description:
-          error instanceof Error ? error.message : "Failed to subscribe",
-        variant: "destructive",
-      });
-    } finally {
-      setLoading(false);
-    }
-  };
+  const { email, setEmail, loading, handleSubmit } =
+    useNewsletterSubscription();
 
   return (
     <div className={`${variant === "small" ? "w-full" : "mx-auto max-w-2xl"}`}>
