+++ src/components/home/Newsletter.tsx
@@ -1,47 +1,11 @@
 "use client";
 
-import { useState } from "react";
-import { useToast } from "~/hooks/use-toast";
+import { useNewsletterSubscription } from "~/hooks/use-newsletter";
 import { cn } from "~/lib/utils";
 
 export function Newsletter() {
-  const [email, setEmail] = useState("");
-  const [loading, setLoading] = useState(false);
-  const { toast } = useToast();
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setLoading(true);
-
-    try {
-      const response = await fetch("/api/subscribe", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ email }),
-      });
-
-      const data = (await response.json()) as { error?: string };
-
-      if (response.ok) {
-        toast({
-          title: "Success!",
-          description: "You've been subscribed to our newsletter.",
-        });
-        setEmail("");
-      } else {
-        throw new Error(data.error ?? "Something went wrong");
-      }
-    } catch (error) {
-      toast({
-        title: "Error",
-        description:
-          error instanceof Error ? error.message : "Failed to subscribe",
-        variant: "destructive",
-      });
-    } finally {
-      setLoading(false);
-    }
-  };
+  const { email, setEmail, loading, handleSubmit } =
+    useNewsletterSubscription();
 
   return (
     <div
