+++ src/components/home/Stats.tsx
@@ -1,10 +1,10 @@
 "use client";
 
-import { useEffect, useState, useRef } from "react";
+import { useEffect, useState } from "react";
 import { Loader } from "../loader";
 import Link from "next/link";
 import { ShareableStatsVideo } from "./ShareableStatsVideo";
-import Confetti from "react-confetti";
+import { Wave } from "./Wave";
 
 interface StatsProps {
   stats: {
@@ -15,123 +15,6 @@ interface StatsProps {
   };
   isLoading?: boolean;
 }
-
-const Wave = ({ fillPercentage }: { fillPercentage: number }) => {
-  const [currentHeight, setCurrentHeight] = useState(0);
-  const [dimensions, setDimensions] = useState({ width: 0, height: 0 });
-  const [showConfetti, setShowConfetti] = useState(false);
-  const containerRef = useRef<HTMLDivElement>(null);
-
-  // Handle container dimensions
-  useEffect(() => {
-    const updateDimensions = () => {
-      if (containerRef.current) {
-        const { width, height } = containerRef.current.getBoundingClientRect();
-        setDimensions({ width, height });
-      }
-    };
-
-    updateDimensions();
-
-    // Create a ResizeObserver to watch for container size changes
-    const resizeObserver = new ResizeObserver(updateDimensions);
-    if (containerRef.current) {
-      resizeObserver.observe(containerRef.current);
-    }
-
-    // Only show confetti after client-side rendering
-    if (fillPercentage === 100) {
-      setShowConfetti(true);
-    } else {
-      setShowConfetti(false);
-    }
-
-    return () => {
-      if (containerRef.current) {
-        resizeObserver.unobserve(containerRef.current);
-      }
-      resizeObserver.disconnect();
-    };
-  }, [fillPercentage]);
-
-  useEffect(() => {
-    // Start at 0 and animate to the target fillPercentage
-    setCurrentHeight(0);
-    setTimeout(() => setCurrentHeight(fillPercentage), 100);
-  }, [fillPercentage]);
-
-  return (
-    <div
-      ref={containerRef}
-      className="absolute bottom-[0px] left-0 z-0 w-full bg-pink transition-all duration-1000"
-      style={{ height: `${currentHeight}%` }}
-    >
-      <div className="absolute bottom-[100%] left-0 z-0 h-[50px] w-full overflow-hidden transition-all duration-1000">
-        <div className="wave-container absolute bottom-0 left-0 w-full">
-          <svg
-            className="waves"
-            xmlns="http://www.w3.org/2000/svg"
-            xmlnsXlink="http://www.w3.org/1999/xlink"
-            viewBox="0 24 150 28"
-            preserveAspectRatio="none"
-            shapeRendering="auto"
-          >
-            <defs>
-              <path
-                id="gentle-wave"
-                d="M-160 44c30 0 58-18 88-18s 58 18 88 18 58-18 88-18 58 18 88 18 v44h-352z"
-              />
-            </defs>
-            <g className="parallax">
-              <use
-                xlinkHref="#gentle-wave"
-                x="48"
-                y="0"
-                fill="rgba(255,0,255,0.7)"
-              />
-              <use
-                xlinkHref="#gentle-wave"
-                x="48"
-                y="3"
-                fill="rgba(255,0,255,0.5)"
-              />
-              <use
-                xlinkHref="#gentle-wave"
-                x="48"
-                y="5"
-                fill="rgba(255,0,255,0.3)"
-              />
-              <use
-                xlinkHref="#gentle-wave"
-                x="48"
-                y="7"
-                fill="rgb(255 0 255)"
-              />
-            </g>
-          </svg>
-        </div>
-      </div>
-      {/* if fillPercentage is 100, show confetti */}
-      {showConfetti && dimensions.width > 0 && dimensions.height > 0 && (
-        <div className="absolute inset-0 z-0 h-full w-full">
-          <Confetti
-            width={dimensions.width}
-            height={dimensions.height}
-            numberOfPieces={150}
-            gravity={0.05}
-            colors={["#eaff00", "#ff03ff", "#180138", "#542299", "#efff40"]}
-            drawShape={(ctx) => {
-              ctx.beginPath();
-              // Draw a small circle
-              ctx.arc(0, 0, 4, 0, 2 * Math.PI);
-              ctx.fill();
-            }}
-          />
-        </div>
-      )}
-    </div>
-  );
-};
 
 export function Stats({ stats, isLoading }: StatsProps) {
   const [displayedPace, setDisplayedPace] = useState(0);
@@ -230,7 +113,7 @@ export function Stats({ stats, isLoading }: StatsProps) {
       <div className="relative flex min-h-[25em] flex-[2] flex-col justify-center overflow-hidden bg-darkpurple-light p-8 md:min-h-[28em]">
         <ShareableStatsVideo stats={stats} />
 
-        <Wave fillPercentage={fillPercentage} />
+        <Wave fillPercentage={fillPercentage} showConfetti />
 
         {/* Add a red dot animated on the bottom left in case there are live games */}
