+++ src/components/standings/StandingsPage.tsx
@@ -5,41 +5,16 @@ import { api } from "~/trpc/react";
 import { StandingsFilters } from "./StandingsFilters";
 import { StandingsGrid } from "./StandingsGrid";
 import { Cta } from "../Cta";
+import { useSeasonFallback } from "~/hooks/use-season-fallback";
 
 export function StandingsPage() {
-  const { data: global, isLoading: isLoadingGlobal } =
-    api.admin.getGlobal.useQuery();
-  const { data: seasons, isLoading: isLoadingSeasons } =
-    api.season.getAllWithStats.useQuery();
-  const { data: stats, isLoading: isLoadingStats } =
-    api.wedgie.getStats.useQuery();
-
-  const defaultSeason = global?.currentSeason?.name ?? "2025/26";
-
-  // Find previous season when current season has 0 wedgies
-  const getPreviousSeason = () => {
-    if (!seasons || !global?.currentSeason?.name) return null;
-
-    const currentSeasonIndex = seasons.findIndex(
-      (s) => s.name === global.currentSeason.name,
-    );
-
-    // If current season is at index 0, there's no previous season
-    // In this case, we should show the most recent season with wedgies
-    if (currentSeasonIndex === 0) {
-      const seasonWithWedgies = seasons.find(
-        (season) => season.totalWedgies > 0,
-      );
-      return seasonWithWedgies ?? null;
-    }
-
-    const previousSeason = seasons[currentSeasonIndex - 1];
-    return previousSeason;
-  };
-
-  const previousSeason = getPreviousSeason();
-  const shouldShowPreviousSeason =
-    stats?.currentSeasonWedgies === 0 && previousSeason;
+  const {
+    global,
+    defaultSeason,
+    previousSeason,
+    shouldShowPreviousSeason,
+    isLoading: isLoadingSeasonData,
+  } = useSeasonFallback();
 
   const [selectedSeason, setSelectedSeason] = useState(defaultSeason);
   const [includeOpponents, setIncludeOpponents] = useState(true);
@@ -51,12 +26,7 @@ export function StandingsPage() {
     } else if (global?.currentSeason?.name) {
       setSelectedSeason(global.currentSeason.name);
     }
-  }, [
-    shouldShowPreviousSeason,
-    previousSeason,
-    global?.currentSeason?.name,
-    stats?.currentSeasonWedgies,
-  ]);
+  }, [shouldShowPreviousSeason, previousSeason, global?.currentSeason?.name]);
 
   const { data: standings, isLoading } = api.wedgie.getSeasonStandings.useQuery(
     {
@@ -66,7 +36,7 @@ export function StandingsPage() {
     { enabled: true },
   );
 
-  if (isLoading || isLoadingGlobal || isLoadingSeasons || isLoadingStats) {
+  if (isLoading || isLoadingSeasonData) {
     return (
       <div className="container mx-auto max-w-3xl text-white">
         <StandingsFilters
@@ -97,7 +67,7 @@ export function StandingsPage() {
       {shouldShowPreviousSeason && (
         <div className="mb-4 rounded-lg border border-pink/30 bg-pink/20 p-4 text-center">
           <p className="text-sm font-bold text-pink">
-            Current season has no wedgies yet. Showing {previousSeason.name}{" "}
+            Current season has no wedgies yet. Showing {previousSeason?.name}{" "}
             season standings.
           </p>
         </div>
